requirements.txt
openai==0.27.0
Pillow==9.5.0
python-dotenv==1.0.0
streamlit==1.20.0
torch==1.13.1
tqdm==4.65.0
transformers==4.27.4
streamlit==1.20.0
